Use the Dockerfile from the main branch
There is no way to change the Dockerfile in a tag. We need to
use the Dockerfile in the main branch. So even if we're buildin
the image for a stable version, we need to check out the main branch.

The Dockerfile will take care of checking out the proper tags.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>