Fix pool edit from pool list
When clicking on the edit button in the pools list, the modal would list
all the pool details except the extra specs. This is expected because when
listing the pools, we omit extra_specs. This can lead to accidentally overwriting
the extra specs for a pool.

This change fetches the entire pool data from the GARM API when clicking
on the edit button in the pools list.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>